+++ cpp/src/io/csv/csv_reader.cu
@@ -29,6 +29,7 @@
 #include <string>
 #include <vector>
 #include <memory>
+#include <unordered_map>
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -607,11 +608,11 @@ gdf_error read_csv(csv_read_arg *args)
 		if (error != GDF_SUCCESS) {
 			return error;
 		}
-		const int h_num_cols = raw_csv.col_names.size();
+		raw_csv.num_actual_cols = raw_csv.num_active_cols = raw_csv.col_names.size();
 
 		// Initialize a boolean array that states if a column needs to read or filtered.
-		raw_csv.h_parseCol = thrust::host_vector<bool>(h_num_cols, true);
-		
+		raw_csv.h_parseCol = thrust::host_vector<bool>(raw_csv.num_actual_cols, true);
+
 		// Rename empty column names to "Unnamed: col_index"
 		for (size_t col_idx = 0; col_idx < raw_csv.col_names.size(); ++col_idx) {
 			if (raw_csv.col_names[col_idx].empty()) {
@@ -619,40 +620,27 @@ gdf_error read_csv(csv_read_arg *args)
 			}
 		}
 
-		int h_dup_cols_removed = 0;
 		// Looking for duplicates
-		for (auto it = raw_csv.col_names.begin(); it != raw_csv.col_names.end(); it++){
-			bool found_dupe = false;
-			for (auto it2 = (it+1); it2 != raw_csv.col_names.end(); it2++){
-				if (*it==*it2){
-					found_dupe=true;
-					break;
+		std::unordered_map<string, int> col_names_histogram;
+		for (auto& col_name: raw_csv.col_names){
+			// Operator [] inserts a default-initialized value if the given key is not present
+			if (++col_names_histogram[col_name] > 1){
+				if (args->mangle_dupe_cols) {
+					// Rename duplicates of column X as X.1, X.2, ...; First appearance stays as X
+					col_name += "." + std::to_string(col_names_histogram[col_name] - 1);
 				}
-			}
-			if(found_dupe){
-				int count=1;
-				for (auto it2 = (it+1); it2 != raw_csv.col_names.end(); it2++){
-					if (*it==*it2){
-						if(args->mangle_dupe_cols){
-							// Replace all the duplicates of column X with X.1,X.2,... First appearance stays as X.
-							std::string newColName  = *it2;
-							newColName += "." + std::to_string(count); 
-							count++;
-							*it2 = newColName;							
-						} else{
-							// All duplicate fields will be ignored.
-							int pos=std::distance(raw_csv.col_names.begin(), it2);
-							raw_csv.h_parseCol[pos]=false;
-							h_dup_cols_removed++;
-						}
-					}
+				else {
+					// All duplicate columns will be ignored; First appearance is parsed
+					const auto idx = &col_name - raw_csv.col_names.data();
+					raw_csv.h_parseCol[idx] = false;
 				}
 			}
 		}
 
-		raw_csv.num_actual_cols = h_num_cols;							// Actual number of columns in the CSV file
-		raw_csv.num_active_cols = h_num_cols-h_dup_cols_removed;		// Number of fields that need to be processed based on duplicatation fields
-
+		// Update the number of columns to be processed, if some might have been removed
+		if (!args->mangle_dupe_cols) {
+			raw_csv.num_active_cols = col_names_histogram.size();
+		}
 	}
 	else {
 		raw_csv.h_parseCol = thrust::host_vector<bool>(args->num_cols, true);
@@ -1266,6 +1254,7 @@ void dataTypeDetection(char *raw_csv,
 			long countDecimal=0;
 			long countSlash=0;
 			long countDash=0;
+			long countPlus=0;
 			long countColon=0;
 			long countString=0;
 			long countExponent=0;
@@ -1290,6 +1279,8 @@ void dataTypeDetection(char *raw_csv,
 						countDecimal++;break;
 					case '-':
 						countDash++; break;
+					case '+':
+						countPlus++; break;
 					case '/':
 						countSlash++;break;
 					case ':':
@@ -1303,11 +1294,12 @@ void dataTypeDetection(char *raw_csv,
 						break;	
 				}
 			}
+			const int countSign = countDash + countPlus;
 
 			// Integers have to have the length of the string
 			long int_req_number_cnt = strLen;
 			// Off by one if they start with a minus sign
-			if(raw_csv[start]=='-' && strLen > 1){
+			if((raw_csv[start]=='-' || raw_csv[start]=='+') && strLen > 1){
 				--int_req_number_cnt;
 			}
 			// Off by one if they are a hexadecimal number
@@ -1340,7 +1332,7 @@ void dataTypeDetection(char *raw_csv,
 					atomicAdd(& d_columnData[actual_col].countInt8, 1L);
 				}
 			}
-			else if(isLikeFloat(strLen, countNumber, countDecimal, countDash, countExponent)){
+			else if(isLikeFloat(strLen, countNumber, countDecimal, countSign, countExponent)){
 					atomicAdd(& d_columnData[actual_col].countFloat, 1L);
 			}
 			// The date-time field cannot have more than 3 strings. As such if an entry has more than 3 string characters, it is not 
@@ -1351,17 +1343,17 @@ void dataTypeDetection(char *raw_csv,
 			else {
 				// A date field can have either one or two '-' or '\'. A legal combination will only have one of them.
 				// To simplify the process of auto column detection, we are not covering all the date-time formation permutations.
-				if((countDash>0 && countDash<=2 && countSlash==0)|| (countDash==0 && countSlash>0 && 	countSlash<=2) ){
+				if((countDash>0 && countDash<=2 && countSlash==0)|| (countDash==0 && countSlash>0 && countSlash<=2) ){
 					if((countColon<=2)){
 						atomicAdd(& d_columnData[actual_col].countDateAndTime, 1L);
 					}
 					else{
-						atomicAdd(& d_columnData[actual_col].countString, 1L);					
+						atomicAdd(& d_columnData[actual_col].countString, 1L);
 					}
 				}
 				// Default field is string type.
 				else{
-					atomicAdd(& d_columnData[actual_col].countString, 1L);					
+					atomicAdd(& d_columnData[actual_col].countString, 1L);
 				}
 			}
 			actual_col++;